t0lte, i9305: Disable samsung_extdisp driver
This is based on commit 6571101 from AGNi kernel.

The samsung_extdisp driver for GT-I9500 was affected by CVE-2015-1800 and CVE-2015-1801. The
Exynos 4 uses an older version of this driver and is at least affected
by CVE-2015-1801. Newer kernel versions for the GT-I9500 had the driver
completely removed. There is no indication that the driver is actually
needed, so let's disable it.

vulnerability disclosure and further information:
http://blog.quarkslab.com/kernel-vulnerabilities-in-the-samsung-s4.html